% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{checkDataFrame}
\alias{checkDataFrame}
\title{Check if the data frame is according the expected package
specification.}
\usage{
checkDataFrame(data)
}
\arguments{
\item{data}{The data frame to be checked}
}
\value{
TRUE for a correct data frame and FALSE for a
wrong data frame.
}
\description{
Check if the data frame is according the expected package
specification.
}
\examples{
\dontrun{
checkDataFrame()
}
}
