% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{compareCityPlot}
\alias{compareCityPlot}
\title{Plot to compare the cases growth in two cities}
\usage{
compareCityPlot(data, type)
}
\arguments{
\item{data}{A data frame generated by the compareTrend function}

\item{type}{The type of plot column or line}
}
\value{
A ggplot to compare two cities
}
\description{
Plot to compare the cases growth in two cities
}
\examples{
\dontrun{
compareCityPlot()
}
}
