% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{compareTrend}
\alias{compareTrend}
\title{Get data to compare two towns by days since case n}
\usage{
compareTrend(data, town1, town2, cases_threshold)
}
\arguments{
\item{data}{A data frame obtained via the daysSinceCase() function}

\item{town1}{The first municipality}

\item{town2}{The second municipality}

\item{cases_threshold}{The number of cases threshold to compare}
}
\value{
A data frame with two municipalities
}
\description{
Get data to compare two towns by days since case n
}
\examples{
\dontrun{
compareTrend()
}
}
