% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{daysSinceCase}
\alias{daysSinceCase}
\title{Calculate days since first case}
\usage{
daysSinceCase(data, cases_threshold)
}
\arguments{
\item{data}{Data frame with city, report_date and cases}

\item{cases_threshold}{TThe number of cases threshold to compare}
}
\value{
Data frame to compare days since case.
}
\description{
Calculate days since first case
}
\examples{
\dontrun{
daysSinceCase()
}
}
