% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{filterTown}
\alias{filterTown}
\title{Filter on municipality}
\usage{
filterTown(data, town)
}
\arguments{
\item{data}{A data frame accorindg the package standards}

\item{town}{The name of the municipality on which the filter
should take place.}
}
\value{
A data frame based on the selected municipality in
the town argument.
}
\description{
Filter on municipality
}
\examples{
\dontrun{
filterTown()
}
}
