% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{loadDataCOVID19NL}
\alias{loadDataCOVID19NL}
\title{Import data from COVID-19-NL Github Repository}
\usage{
loadDataCOVID19NL(
  fromGitHub = TRUE,
 
    path = "https://raw.githubusercontent.com/ZequnZ/COVID-19-NL/master/data/NL_COVID19_info_city.csv"
)
}
\arguments{
\item{fromGitHub}{If this parameter is TRUE the data will be loaded
from Github. If FALSE the a local csv file is expected.}

\item{path}{Mainly to be used with the fromGitHub == FALSE
option. Point to the file `NL_COVID19_info_city.csv`. Standard path is
`https://raw.githubusercontent.com/ZequnZ/COVID-19-NL/master/data/NL_COVID19_info_city.csv`}
}
\value{
A data frame grouped with reported cases per municipality and day.
}
\description{
Import data from COVID-19-NL Github Repository
}
\examples{
\dontrun{
loadDataCOVID19NL()
}
}
