% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCases}
\alias{plotCases}
\title{Plots the daily positive tested cases.}
\usage{
plotCases(data, town, type, trend_line)
}
\arguments{
\item{data}{Data frame container the case data. Should be
in the expected package standard. Which can be checked by
the `checkDataFrame()` function.}

\item{town}{Municipality for whoch the plot should be
created. If no town is specified the sum of all
municipalities is used.}

\item{type}{The type op plot. Inputs can be `columns`,
`line` or `both`.}

\item{trend_line}{If `trend_line == TRUE` a trendline will
be added to the plot.}
}
\value{
A ggplot and a dataframe based on the municipality
in the input. The plot can be called by `$plot` and the
data via `$data`.
}
\description{
Plots the daily positive tested cases.
}
\examples{
\dontrun{
plotCases(md, town = "Amsterdam", type = "column", trend_line = TRUE)
}
}
