% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prettyDeltaCases}
\alias{prettyDeltaCases}
\title{Pretty delta cases format for plots}
\usage{
prettyDeltaCases(x)
}
\arguments{
\item{x}{The value to make pretty}
}
\value{
The imput value in a + or - minus format
}
\description{
Pretty delta cases format for plots
}
\examples{
prettyDeltaCases(x = 10)
prettyDeltaCases(x = 0)
prettyDeltaCases(x = -10)
}
