% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{loadDataCOVID19NL}
\alias{loadDataCOVID19NL}
\title{Import data from COVID-19-NL Github Repository}
\usage{
loadDataCOVID19NL(path)
}
\arguments{
\item{path}{The path the file `NL_COVID19_info_city.csv`.}
}
\value{
A data frame grouped with reported cases per municipality and day.
}
\description{
Import data from COVID-19-NL Github Repository
}
\examples{
\dontrun{
loadDataCOVID19NL()
}
}
